requirements.txt
# Django framework
Django==4.2.28

# REST API & JWT
djangorestframework==3.20.0
djangorestframework-simplejwt==6.1.0

# Database (SQLite is default, but for PostgreSQL if needed)
psycopg2-binary==2.9.7

# Face Recognition & Liveness
opencv-python==4.7.0.72
mediapipe==0.10.32
face-recognition==1.3.0
numpy==1.26.5

# Image handling
Pillow==10.1.0

# PDF generation for election results
reportlab==4.1.10

# Charts & visualization (optional, if using JS charts, can be skipped)
matplotlib==3.9.3

# Security & cryptography (for vote encryption)
cryptography==41.0.3

# Environment management 
python-dotenv==1.0.0

# Required Package
python-dotenv